rustPlatform: fix cargoSetupHook for cross-compiling
this patch changes cargoSetupHook to use targetPlatform instead
of hostPlatform when checking for a target package's build
dependencies. Since cargoSetupHook is a nativeBuildInput, when it
compares platforms to check it it is running a cross compile
it needs to use targetPlatform rather than hostPlatform else
only host tooling will be used.

fixes #355803